+++ test/jiffy_display_test.dart
@@ -4,14 +4,12 @@ import 'package:test/test.dart';
 void main() {
   group('Test format and default format datetime methods', () {
     test("Test format method", () {
-      expect(Jiffy("2019, 10, 16", "yyyy, MM, dd").format("MMMM dd, yyyy"),
-          "October 16, 2019");
-      expect(Jiffy("2019, 10, 16", "yyyy, MM, dd").format("'Today is' dd MMM"),
+      expect(Jiffy("20191016").format("MMMM dd, yyyy"), "October 16, 2019");
+      expect(Jiffy("2019-10-16 12:00").format("'Today is' dd MMM"),
           "Today is 16 Oct");
-      expect(Jiffy("2019, 10, 16", "yyyy, MM, dd").format("MMMMEEEEd"),
+      expect(Jiffy("2019-10-16T00:00:00.000").format("MMMMEEEEd"),
           "Wednesday, October 16");
-      expect(Jiffy("2019, 10, 16", "yyyy, MM, dd").format(),
-          "2019-10-16T00:00:00.000");
+      expect(Jiffy("2019/10/16").format(), "2019-10-16T00:00:00.000");
     });
     test("Test from method", () {
       expect(Jiffy([2019, 10, 16]).E, "Wed");